backend_ppe/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
pillow>=10.3.0
numpy<2.0.0
torch>=2.1.0
torchvision>=0.16.0
